# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_UniverseHappiness_WiseDx_internal_errors.AppError:
    properties:
      code:
        $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_UniverseHappiness_WiseDx_internal_errors.ErrorCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 2000
    - 2001
    - 2002
    - 2003
    - 2004
    - 2100
    - 2101
    - 2102
    - 2103
    type: integer
    x-enum-varnames:
    - ErrBadRequest
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrMethodNotAllowed
    - ErrConflict
    - ErrTooManyRequests
    - ErrInternalServer
    - ErrServiceUnavailable
    - ErrTimeout
    - ErrValidation
    - ErrTenantNotFound
    - ErrTenantAlreadyExists
    - ErrTenantInactive
    - ErrTenantNameRequired
    - ErrTenantInvalidStatus
    - ErrAgentMissingThinkingModel
    - ErrAgentMissingAllowedTools
    - ErrAgentInvalidMaxIterations
    - ErrAgentInvalidTemperature
...
  internal_handler_session.CreateKnowledgeQARequest:
    properties:
      agent_enabled:
        description: Whether agent mode is enabled for this request
        type: boolean
      knowledge_base_ids:
        description: Selected knowledge base ID for this request
        items:
          type: string
        type: array
      knowledge_ids:
        description: Selected knowledge ID for this request
        items:
          type: string
        type: array
      mentioned_items:
        description: '@mentioned knowledge bases and files'
        items:
          $ref: '#/definitions/internal_handler_session.MentionedItemRequest'
        type: array
      query:
        description: Query text for knowledge base search
        type: string
      summary_model_id:
        description: Optional summary model ID for this request (overrides session
          default)
        type: string
      web_search_enabled:
        description: Whether web search is enabled for this request
        type: boolean
    required:
    - query
    type: object
...
  internal_handler_session.MentionedItemRequest:
    properties:
      id:
        type: string
      kb_type:
        description: '"document" or "faq" (only for kb type)'
        type: string
      name:
        type: string
      type:
        description: '"kb" for knowledge base, "file" for file'
        type: string
    type: object
  internal_handler_session.SearchKnowledgeRequest:
    properties:
      knowledge_base_id:
        description: Single knowledge base ID (for backward compatibility)
        type: string
      knowledge_base_ids:
        description: IDs of knowledge bases to search (multi-KB support)
        items:
          type: string
        type: array
      knowledge_ids:
        description: IDs of specific knowledge (files) to search
        items:
          type: string
        type: array
      query:
        description: Query text to search for
        type: string
    required:
    - query
    type: object
...
info:
  contact:
    name: WeKnora Github
    url: https://github.com/UniverseHappiness/WiseDx
  description: WeKnora 知识库管理系统 API 文档
  termsOfService: http://swagger.io/terms/
  title: WeKnora API
  version: "1.0"
paths:
...
  /sessions/{session_id}/agent-qa:
    post:
      consumes:
      - application/json
      description: 基于Agent的智能问答，支持多轮对话和SSE流式响应
      parameters:
      - description: 会话ID
        in: path
        name: session_id
        required: true
        type: string
      - description: 问答请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_session.CreateKnowledgeQARequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: 问答结果（SSE流）
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: Agent问答
      tags:
      - 问答
...
  /sessions/{session_id}/knowledge-qa:
    post:
      consumes:
      - application/json
      description: 基于知识库的问答（使用LLM总结），支持SSE流式响应
      parameters:
      - description: 会话ID
        in: path
        name: session_id
        required: true
        type: string
      - description: 问答请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_session.CreateKnowledgeQARequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: 问答结果（SSE流）
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 知识问答
      tags:
      - 问答
...
  /sessions/search:
    post:
      consumes:
      - application/json
      description: 在知识库中搜索（不使用LLM总结）
      parameters:
      - description: 搜索请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_session.SearchKnowledgeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 搜索结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_UniverseHappiness_WiseDx_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 知识搜索
      tags:
      - 问答
...
securityDefinitions:
  ApiKeyAuth:
    description: 租户身份认证：输入 sk- 开头的 API Key
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: 用户登录认证：输入 Bearer {token} 格式的 JWT 令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"